migrate to wallet component
