+++ src/App.tsx
@@ -7,6 +7,7 @@ import { Providers, connector } from './config/web3'
 import { Home } from './views/Home'
 import { ProposalContractABI, PROPOSAL_CONTRACT_ADDRESS_GOERLI } from './config/artifacts'
 import { Header } from './components/Header'
+import { Switch } from "./components/Switch"
 
 
 type Library = {
@@ -52,38 +53,6 @@ function App() {
   const { active, error, activate, deactivate, account, library, chainId } = useWeb3React<Providers>()
   const [ selectedLibrary, setSelectedLibrary ] = useState(library?.web3)
   const [ switchLibrary, setSwitchLibrary] = useState(true)
-  const [ balance, setBalance ] = useState(0)
-
-  const connect = useCallback(()  => {
-    localStorage.setItem('previouslyConnected', 'true')
-    activate(connector)
-  }, [activate])
-
-  const getBalance = useCallback(async () => {
-    let balance : number;
-    balance = (await selectedLibrary?.getBalance(account!!)) as number ?? 0 
-    balance = Number(( balance / 1e18).toFixed(2));
-    setBalance(balance)
-  }, [ selectedLibrary, account ])
-
-
-  const disconnect = () => {
-    deactivate()
-    localStorage.removeItem('previouslyConnected')
-  }
-  const isUnsupportedChain = error instanceof UnsupportedChainIdError 
-
-  useEffect(() => {
-    if (active) {
-      getBalance()
-    }
-  }, [active, getBalance])
-
-  useEffect(() => {
-    if (localStorage.getItem('previouslyConnected') === 'true') {
-      connect()
-    }
-  }, [connect])
 
   useEffect(() => {
     doSwitchLibrary(switchLibrary,setSelectedLibrary, library)
@@ -102,31 +71,9 @@ function App() {
     <div className="App">
       <LibraryContext.Provider value={{active, account, activate, error, deactivate, chainId, selectedLibrary,setSelectedLibrary}}>
       <Header />
-      <input 
-        checked={switchLibrary}
-        type="checkbox"
-        onChange={() => { setSwitchLibrary(!switchLibrary) }}
-      />
       <header className="App-header">
+        <Switch switchLibrary={switchLibrary} setSwitchLibrary={setSwitchLibrary} />
         <Home />
-        {
-          active ?
-            <div>
-              <p>{account} - {balance}</p>
-              <button onClick={disconnect}> Desconectar </button>
-            </div>
-          : 
-          isUnsupportedChain ? 
-            'Red no soportada'
-            :
-            Object.keys(connectorsByName).map((name) => {
-              return (
-                <button key={name} onClick={connect}>
-                 Connect
-                </button>
-              )
-            })
-        }
       </header>
       </LibraryContext.Provider>
     </div>
